Changed User's id to a long.

+++ backend/jobapplicationbackend/src/main/java/com/jpettit/jobapplicationbackend/models/user/User.java
@@ -2,10 +2,7 @@ package com.jpettit.jobapplicationbackend.models.user;
 
 import com.jpettit.jobapplicationbackend.enums.Role;
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -22,7 +19,7 @@ import java.util.List;
 public class User implements UserDetails {
     @Id
     @GeneratedValue
-    private Integer id;
+    private Long id;
     private String firstname;
     private String lastname;
     private String email;
